delete plot_helper.jl
